moved login_required into journalist_app/decorators.py
