+++ install_files/ansible-base/roles/build-securedrop-app-code-deb-pkg/files/usr.sbin.apache2
@@ -132,6 +132,8 @@
   /var/www/securedrop/journalist_app/__init__.py r,
   /var/www/securedrop/journalist_app/__init__.pyc rw,
   /var/www/securedrop/journalist_app/__pycache__.py rw,
+  /var/www/securedrop/journalist_app/decorators.py r,
+  /var/www/securedrop/journalist_app/decorators.pyc rw,
   /var/www/securedrop/journalist_app/utils.py r,
   /var/www/securedrop/journalist_app/utils.pyc rw,
   /var/www/securedrop/journalist_templates/_source_row.html r,
